index: move DefaultIndexStore to new default_index_store.rs
