+++ lib/src/lib.rs
@@ -22,6 +22,7 @@ pub mod commit;
 pub mod commit_builder;
 pub mod conflicts;
 pub mod dag_walk;
+mod default_index_store;
 pub mod diff;
 pub mod file_util;
 pub mod files;
